ASIN関連APIテストをリファクタ
