+++ test/adapters/amazonAdapter.test.ts
@@ -1,9 +1,7 @@
 import { expect, test, vi } from "vitest"
-import { 
+import {
   createKVCacheAdapter,
-  createSlackLoggerAdapter,
-  type CacheAdapter,
-  type LoggerAdapter
+  createSlackLoggerAdapter
 } from "../../functions/src/adapters/amazonAdapter"
 import { createMockAmazonResponse } from "../helpers/mockData"
 
@@ -20,22 +18,28 @@ test("KV cache adapter works correctly", async () => {
     list: vi.fn().mockResolvedValue({ keys: [], list_complete: true }),
     getWithMetadata: vi.fn().mockResolvedValue({ value: null, metadata: null })
   } as unknown as KVNamespace
-  
+
   const cache = createKVCacheAdapter(mockKV)
-  
+
   await cache.get("test-key")
   expect(mockKV.get).toHaveBeenCalledWith("test-key", "json")
-  
+
   const testData = createMockAmazonResponse("TEST123")
   await cache.put("test-key", testData, 3600)
-  expect(mockKV.put).toHaveBeenCalledWith("test-key", JSON.stringify(testData), {
-    expirationTtl: 3600
-  })
+  expect(mockKV.put).toHaveBeenCalledWith(
+    "test-key",
+    JSON.stringify(testData),
+    {
+      expirationTtl: 3600
+    }
+  )
 })
 
 test("Slack logger adapter works correctly", async () => {
   const logger = createSlackLoggerAdapter("https://hooks.slack.com/test")
-  
+
   // The actual postLogToSlack is mocked, so this should not throw
-  await expect(logger.logError("Test message", "http://test.com")).resolves.toBeUndefined()
+  await expect(
+    logger.logError("Test message", "http://test.com")
+  ).resolves.toBeUndefined()
 })
